requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
openai==1.47.0
requests==2.32.3
beautifulsoup4==4.12.3
gTTS==2.5.3
Jinja2==3.1.4
moviepy==1.0.3
playwright==1.48.0
